# lib/api-zod/src/generated/types/createCaseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LexCom API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCaseBody {
  title: string;
  description?: string;
  caseType: string;
  hearingDate?: Date | null;
}